17_sites_monitoring/requirements.txt
futures==3.1.1
python-dateutil==2.6.1
python-whois==0.6.5
requests==2.18.4
